Duplicate state logic to reducer for pessimistic UI portion of assignment
